runtime: extract Windows PE globals scan helper
